chore: adds path responses
