+++ sample/plugins/gatsby-source-openapi-aggregate/processors/factory.js
@@ -1,0 +1,11 @@
+const spec20Processor = require('./spec20Processor')
+
+const specProcessorFactory = spec => {
+  if (spec.swagger === '2.0') {
+    return spec20Processor
+  }
+
+  throw new Error(`Unsupported spec ${spec.swagger}`)
+}
+
+module.exports = specProcessorFactory
